bagfix

+++ src/main/java/com/example/whats_happened/service/IncidentService.java
@@ -22,12 +22,10 @@ public class IncidentService {
     }
 
     public List<Incident> getAllIncidents(){
-        log.info("All incidents received");
         return incidentRepository.findAll();
     }
 
     public Incident getIncidentById(UUID id){
-        log.info("Incident received with id=" + id);
         return incidentRepository.findById(id).orElse(null);
     }
 
@@ -35,18 +33,13 @@ public class IncidentService {
         if(incident.getCreatedAt() == null){
             incident.setCreatedAt(LocalDateTime.now());
         }
-        log.info("Incident saved, id=" + incident.getId());
-        return incidentRepository.save(incident);
-    }
-
-    public Incident updateIncident(Incident incident){
-        log.info("Incident updated with id=" + incident.getId());
         return incidentRepository.save(incident);
     }
 
     public void deleteIncidentById(UUID id){
-        log.info("Incident deleted with id=" + id);
-        incidentRepository.deleteById(id);
+        if(incidentRepository.findById(id).isPresent()){
+            incidentRepository.deleteById(id);
+        }
     }
 
 }
